+++ phx_installer/test/openmaize_phx_test.exs
@@ -32,8 +32,9 @@ defmodule Mix.Tasks.Openmaize.PhxTest do
         assert file =~ ~s(resources "/users", UserController)
       end
 
-      assert_received {:mix_shell, :info, ["\nPlease check the generated" <> _ = message]}
-      assert message =~ ~s(See the documentation for Openmaize.Config)
+      assert_received {:mix_shell, :info, ["\nWe're almost done!" <> _ = message]}
+      assert message =~ ~s({:openmaize, {"~> 2.7"}})
+      assert message =~ ~s(After that, run `mix test` to run all the tests)
     end
   end
 
@@ -113,12 +114,17 @@ defmodule Mix.Tasks.Openmaize.PhxTest do
   end
 
   test "generates resource with email unique_id" do
-    in_tmp "generates default html resource", fn ->
+    in_tmp "generates resource with email unique_id", fn ->
       Mix.Tasks.Openmaize.Phx.run ["email"]
 
       assert_file "web/controllers/session_controller.ex", fn file ->
         assert file =~ "plug Openmaize.Login, [unique_id: :email] when action in [:create]"
         refute file =~ "def confirm_email(%Plug.Conn{private: %{openmaize_error: message}}"
+      end
+
+      assert_file "test/controllers/session_controller_test.exs", fn file ->
+        assert file =~ ~s(@valid_attrs %{email: "robin@mail.com")
+        assert file =~ ~s(@invalid_attrs %{email: "robin@mail.com")
       end
 
       assert_file "web/models/user.ex", fn file ->
@@ -130,12 +136,17 @@ defmodule Mix.Tasks.Openmaize.PhxTest do
   end
 
   test "generates resource with custom unique_id" do
-    in_tmp "generates default html resource", fn ->
+    in_tmp "generates resource with custom unique_id", fn ->
       Mix.Tasks.Openmaize.Phx.run ["aaarrgh"]
 
       assert_file "web/controllers/session_controller.ex", fn file ->
         assert file =~ "plug Openmaize.Login, [unique_id: :aaarrgh] when action in [:create]"
         refute file =~ "def confirm_email(%Plug.Conn{private: %{openmaize_error: message}}"
+      end
+
+      assert_file "test/controllers/session_controller_test.exs", fn file ->
+        assert file =~ ~s(@valid_attrs %{aaarrgh: "robin")
+        assert file =~ ~s(@invalid_attrs %{aaarrgh: "robin")
       end
 
       assert_file "web/models/user.ex", fn file ->